build: add pyproject.toml for project configuration
This commit adds a new file, pyproject.toml, which includes the build system configuration and project metadata. The pyproject.toml file specifies the required package "hatchling" and sets it as the build backend. Additionally, it defines the project name as "api" with a dynamic version, describes it as an API handling backend for a personal website, and lists the authors' information. The requires-python field is set to ">=3.11.0", and the dependencies section includes packages such as APScheduler 3.10.4, Django 4.2.4, django_apscheduler 0.6.2, djangorestframework 3.14.0, and Requests 2.31.0 that are necessary for the project's functionality.