fix(api): add response import and update arc_metric_api function
This commit adds the import statement for `Response` from `rest_framework.response` module in `api/views.py`. It also updates the `arc_metric_api` function to return either a `JsonResponse` or a `Response` object based on the validation of variables. If variables are not set correctly, it now returns a response with the message "please set variables correctly".